* Rename a number of methods so as to make clearer the distinction between system and application channel counts, rates, etc.  Also change a number of prototypes from using size_t to using int.

+++ JACKAudioIO.h
@@ -31,7 +31,7 @@ public:
     virtual double getCurrentTime() const;
 
 protected:
-    void setup(size_t channels);
+    void setup(int channels);
     int process(jack_nframes_t nframes);
     int xrun();
